refactor: ReceiptModal 유지보수성, 가독성 개선
