+++ src/components/common/modals/receiptModal/ReceiptModal.tsx
@@ -1,29 +1,37 @@
 import { useState } from 'react';
-import styled, { css } from 'styled-components';
-import ReactModal from 'react-modal';
-import { calcRatio, theme } from '@lib/styles/theme';
-import { TItemName } from '@lib/types/TItemName';
-import PerforatedLineTop from './PerforatedLineTop';
-import PerforatedLineBottom from './PerforatedLineBottom';
-import { TInterestRate } from '@lib/types/IInterestRate';
+import styled from 'styled-components';
+import { theme } from '@lib/styles/theme';
 import '../styles.css';
-import ProposalBadge from '@components/common/badges/ProposalBadge';
+import ReactModal from 'react-modal';
 import getHeightByVariant from './getHeightByVariant';
 import DashedBorder from './DashedBorder';
-import FirstRow from './FirstRow';
-import SecondRow from './SecondRow';
-import ThirdRow from './ThirdRow';
 import SubmitButton from './SubmitButton';
+import { IDongil } from '@lib/types/IDongil';
+import TopContent from './contents/TopContent';
+import BottomContent from './contents/BottomContent';
+import CommentContent from './contents/CommentContent';
+import { TReceiptModalVariant } from './TReceiptModalVariant';
+import PerforatedLineTop from './perforatedLines/PerforatedLineTop';
+import PerforatedLineBottom from './perforatedLines/PerforatedLineBottom';
 
-type TVariant = 'contract' | 'proposing' | 'rejected' | 'proposed';
-
-interface QuaternaryModalProps {
+interface QuaternaryModalProps
+  extends Pick<
+      IDongil,
+      | 'createdAt'
+      | 'interestRate'
+      | 'itemName'
+      | 'title'
+      | 'totalPrice'
+      | 'weekPrice'
+      | 'weeks'
+    >,
+    Pick<Partial<IDongil>, 'comment' | 'fileName'> {
   /**
    * 용도를 선택합니다.
    * contract: 자녀 - 새로 돈길 계약하기 / proposing: 자녀 - 대기중인 돈길 (제안중)
    * rejected: 자녀 - 대기중인 돈길 (거절됨) / proposed: 부모 - 제안받은 돈길
    * */
-  variant: TVariant;
+  variant: TReceiptModalVariant;
   /**
    * submit (모달 하단 버튼 클릭) 시 처리될 지스니스 로직을 처리하는 함수 입니다.
    * 모달 하단 버튼이 1개인 경우 onSubmit만 사용합니다.
@@ -31,17 +39,8 @@ interface QuaternaryModalProps {
    * */
   onSubmit: any;
   onExtraSubmit?: any;
-  createdAt: string;
-  interestRate: TInterestRate;
   isMom: boolean;
-  itemName: TItemName;
-  title: string;
-  totalPrice: number;
-  weekPrice: number;
-  weeks: number;
-  comment?: string;
   shouldCloseOnOverlayClick: boolean;
-  fileName?: string;
   isSubmit?: boolean;
 }
 
@@ -125,40 +124,18 @@ function ReceiptModal({
       <Content>
         <DashedBorder variant={variant} />
         <PerforatedLineTop fill={theme.palette.greyScale.white} />
-        <Top variant={variant}>
-          {variant === 'contract' && (
-            <span className="header">계약서 전송 성공!</span>
-          )}
-          {variant === 'proposing' && <ProposalBadge isProposing />}
-          {variant === 'rejected' && <ProposalBadge isProposing />}
-          <span className="body">{title}</span>
-        </Top>
-
-        <Bottom variant={variant}>
-          <FirstRow isMom={isMom} itemName={itemName} />
-          <SecondRow
-            totalPrice={totalPrice}
-            weekPrice={weekPrice}
-            interestRate={interestRate}
-          />
-          <ThirdRow weeks={weeks} createdAt={createdAt} />
-          <SignatureWrapper>
-            {/* <img
-              src={
-                isSubmit
-                  ? fileName
-                  : `https://bankidz-bucket.s3.ap-northeast-2.amazonaws.com/${fileName}`
-              }
-            /> */}
-          </SignatureWrapper>
-        </Bottom>
-
-        {variant === 'rejected' && (
-          <Comment>
-            <div className="header">부모님의 한줄평</div>
-            <div className="body">{comment}</div>
-          </Comment>
-        )}
+        <TopContent variant={variant} title={title} />
+        <BottomContent
+          variant={variant}
+          isMom={isMom}
+          itemName={itemName}
+          totalPrice={totalPrice}
+          weekPrice={weekPrice}
+          interestRate={interestRate}
+          weeks={weeks}
+          createdAt={createdAt}
+        />
+        {variant === 'rejected' && <CommentContent comment={comment!} />}
         <PerforatedLineBottom fill={theme.palette.greyScale.white} />
         <SubmitButton
           variant={variant}
@@ -193,123 +170,6 @@ const Content = styled.div`
   align-items: center;
   width: 100%;
   position: relative;
-`;
-
-const Top = styled.div<{ variant: TVariant }>`
-  ${({ variant }) =>
-    variant === 'contract' &&
-    css`
-      height: 88px;
-    `}
-  ${({ variant }) =>
-    variant === 'proposed' &&
-    css`
-      height: 88px;
-    `}
-  ${({ variant }) =>
-    (variant === 'proposing' || variant === 'rejected') &&
-    css`
-      height: 116px;
-    `}
-  margin: -2px 0; // overlaps 2px
-  background: ${({ theme }) => theme.palette.greyScale.white};
-  width: 100%;
-
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  border-bottom-left-radius: ${({ theme }) => theme.radius.medium};
-  border-bottom-right-radius: ${({ theme }) => theme.radius.medium};
-
-  .header {
-    ${({ theme }) => theme.typo.text.S_14_M};
-    color: ${({ theme }) => theme.palette.greyScale.grey500};
-    margin-bottom: 12px;
-  }
-  .body {
-    margin-top: ${({ variant }) =>
-      variant === 'proposing' || variant === 'rejected' ? '16px' : '0px'};
-    ${({ theme }) => theme.typo.popup.T_24_EB};
-    color: ${({ theme }) => theme.palette.main.yellow400};
-  }
-`;
-
-const Bottom = styled.div<{ variant: TVariant }>`
-  margin-bottom: -2px; // overlaps 2px
-  background: ${({ theme }) => theme.palette.greyScale.white};
-  width: 100%;
-  height: 360px;
-
-  border-top-left-radius: ${({ theme }) => theme.radius.medium};
-  border-top-right-radius: ${({ theme }) => theme.radius.medium};
-  ${({ variant }) =>
-    variant === 'rejected' &&
-    css`
-      border-bottom-left-radius: ${({ theme }) => theme.radius.medium};
-      border-bottom-right-radius: ${({ theme }) => theme.radius.medium};
-    `}
-
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: space-between;
-  position: relative;
-`;
-
-const Comment = styled.div`
-  border-top-left-radius: ${({ theme }) => theme.radius.medium};
-  border-top-right-radius: ${({ theme }) => theme.radius.medium};
-  width: 100%;
-  height: 86px;
-  background: ${({ theme }) => theme.palette.greyScale.white};
-  margin: -2px 0; // overlaps 2px
-
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: flex-start;
-
-  .header {
-    width: 80px;
-    height: 12px;
-    ${({ theme }) => theme.typo.text.S_12_M};
-    color: ${({ theme }) => theme.palette.greyScale.grey500};
-    margin-left: 24px;
-    margin-top: 18px;
-  }
-  .body {
-    width: 276px;
-    height: 14px;
-    ${({ theme }) => theme.typo.text.T_16_EB}
-    color: ${({ theme }) => theme.palette.sementic.red300};
-    margin-left: 24px;
-    margin-top: 18px;
-    margin-bottom: 32px;
-  }
-`;
-
-const SignatureWrapper = styled.div`
-  z-index: 710;
-  position: absolute;
-  width: ${calcRatio(146, 324)};
-  // TODO: 도영이는 뒤에 글씨 가리는게 별로라고 생각해서 기디 회의 후에
-  // 서명 크기 재조정 필요할것 같습니다.
-  height: 173px;
-  right: 2px;
-  bottom: 0;
-  background: rgba(36, 39, 41, 0.7);
-
-  & > img {
-    max-width: 100%;
-    margin-top: auto;
-    margin-bottom: 16px;
-    height: 120px;
-  }
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
 `;
 
 // http://jsfiddle.net/dineshranawat/Ls95n95L/
