+++ src/components/common/modals/receiptModal/TReceiptModalVariant.ts
@@ -1,0 +1,1 @@
+export type TReceiptModalVariant = 'contract' | 'proposing' | 'rejected' | 'proposed';
