+++ src/components/common/modals/receiptModal/perforatedLines/PerforatedLineBottom.tsx
