+++ src/components/common/modals/receiptModal/perforatedLines/PerforatedLineTop.tsx
